// src/app/features/browse/models/recipe.model.ts
export interface Recipe {
  id: string;
  title: string;
  imageUrl: string;
  api: string;
};

export interface RecipeDetail {
  id: string;
  title: string;
  ingredients: Ingredients[];
  instructions: string;
  totalTime: number;
  servings: number;
  imageUrl: string;
  api: string;
  sourceName: string,
  sourceUrl: string,
}

export interface Ingredients {
  name: string;
  quantity: number;
  unit: string;
}